feat: endpoints definition and codec for sharing
