+++ src/definitions/index.ts
@@ -4,3 +4,4 @@ export * from './location.definition';
 export * from './place.definition';
 export * from './recurrence.definition';
 export * from './persistence.definitions';
+export * from './endpoints.definitions';
